Fixed Samsung's bad code, so AOSP can compile it

+++ drivers/media/radio/s610/radio-s610.c
@@ -1650,7 +1650,7 @@ static int s610_radio_fops_open(struct file *file)
 			}
 
 			abox_request_cpu_gear_sync(radio->a_dev,
-				dev_get_drvdata(radio->a_dev), (unsigned int)radio->dev, ABOX_CPU_GEAR_MAX, "FM");
+				dev_get_drvdata(radio->a_dev), (uintptr_t)radio->dev, ABOX_CPU_GEAR_MAX, "FM");
 		}
 #endif /* USE_AUDIO_PM */
 
@@ -1768,7 +1768,7 @@ static int s610_radio_fops_release(struct file *file)
 #ifdef USE_AUDIO_PM
 		if (radio->a_dev) {
 			abox_request_cpu_gear_sync(radio->a_dev,
-				dev_get_drvdata(radio->a_dev), (unsigned int)radio->dev, ABOX_CPU_GEAR_MIN, "FM");
+				dev_get_drvdata(radio->a_dev), (uintptr_t)radio->dev, ABOX_CPU_GEAR_MIN, "FM");
 
 			pm_runtime_put_sync(radio->a_dev);
 		}
